change np to c_int

+++ src/ppidd.cpp
@@ -615,11 +615,11 @@ extern "C" {
  *  - \b GA analogous to http://hpc.pnl.gov/globalarrays/api/c_op_api.html#LOCATE_REGION
  */
    int PPIDD_Location(int64_t *handle,     /*!< array handle */
-                       int64_t *ilo,        /*!< lower element subscript, 1 (not 0) for first element */
-                       int64_t *ihi,        /*!< higher element subscript */
-                       int64_t *map,        /*!< return start/end index for <tt>proclist</tt> */
-                       int64_t *proclist,   /*!< proc id list */
-                       int64_t *np          /*!< proc number */
+                      int64_t *ilo,        /*!< lower element subscript, 1 (not 0) for first element */
+                      int64_t *ihi,        /*!< higher element subscript */
+                      int64_t *map,        /*!< return start/end index for <tt>proclist</tt> */
+                      int64_t *proclist,   /*!< proc id list */
+                      int *np              /*!< proc number */
    ) {
     switch (ppidd_impl) {
 #ifdef HAVE_MPI_H
